requirements.txt
argparse
logging
os
csv
pyzbar
Pillow
